requirements.txt
telethon
python-dotenv
fastapi
uvicorn
python-multipart